Moved in the new architecture
